Big fixes; Added the try-else statement; InConsoleMode: multiple line supported; process module

+++ src/errors/raiseError.ts
@@ -2,6 +2,10 @@ import { ICM } from "../interpreter/interpreter.js"
 import CompilerError from "./_error.js"
 import UnknowError from "./compiler/UnknowError.js"
 
+export let allowErrors = true
+export function setAllowErrors(allow: boolean) {
+    allowErrors = allow
+}
 export default class RaiseFlyLangCompilerError {
     error: CompilerError
     constructor(error?: CompilerError) {
@@ -9,9 +13,10 @@ export default class RaiseFlyLangCompilerError {
     }
     
     raise() {
+        if(!allowErrors) return this.error
         console.error(this.error.toString())
         
         if(!ICM()) process.kill(process.pid) // This kill the current process (without any additionnal message except the error's one)
-        return null // For typing (when the code above is executed, the process is killed and stop)
+        return this.error // For typing
     }
 }
